Commit 15.04.2020

+++ rms.cpp
@@ -2,7 +2,7 @@
 using namespace std; 
   
 // Functia care calculeaza RMS (Root Mean Square)
-float rmsValue(int arr[], int n) 
+float rmsValue(float arr[],  float n) 
 { 
     int square = 0; 
     float mean = 0.0, root = 0.0; 
@@ -24,14 +24,14 @@ float rmsValue(int arr[], int n)
 // Apelarea functiei RMS 
 int main() 
 { 
-    int arr[] = {37.035, 0.0933,
+   float arr[] = {37.035, 0.0933,
 37.0352, 0.0950,
 37.0355, 0.0783,
 37.0358, 0.0767,
 37.036, 0.0633,
 37.0363, 0.0633,
 37.0365, 0.0283 }; 
-    int n = sizeof(arr) / sizeof(arr[0]); 
+    float n = sizeof(arr) / sizeof(arr[0]); 
   
     cout <<"Valoarea RMS este " << rmsValue(arr, n);
